fix: fix change quantity error & seperate Cart
