+++ client/src/components/Carts/index.js
@@ -1,0 +1,3 @@
+export { default as CartList } from './CartList';
+export { default as CartItem } from './CartItem';
+export { default as OrderHistory } from './OrderHistory';
